move eval list test into new package
